deposit function for storage provider

+++ src/CosmicFil.sol
@@ -4,14 +4,13 @@ pragma solidity ^0.8.13;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract CosmicFil is ERC20 {
-  constructor(string memory name, string memory symbol) ERC20(name, symbol) {
-  }
+    constructor(string memory name, string memory symbol) ERC20(name, symbol) {}
 
-  function mint(address to, uint256 amount) public virtual {
-    _mint(to, amount);
-  }
+    function mint(address to, uint256 amount) public virtual {
+        _mint(to, amount);
+    }
 
-  function burn(address from, uint amount) public virtual {
-    _burn(from, amount);
-  }
+    function burn(address from, uint256 amount) public virtual {
+        _burn(from, amount);
+    }
 }
